pyproject.toml
[project]
name = "fast-ecommerce"
version = "0.2.1"
description = "Ecommerce API"
authors = [
    {name = "Jônatas Oliveira", email = "contact@jonatasoliveira.dev"},
]
license = "AGPL-3.0-or-later"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "aiofiles>=24.1.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.2",
    "asyncpg>=0.29.0",
    "bcrypt>=4.2.0",
    "boto3>=1.35.15",
    "cython>=3.0.11",
    "dynaconf>=3.2.6",
    "fastapi>=0.114.2",
    "faststream[rabbit]>=0.5.23",
    "httpx>=0.27.2",
    "jinja2>=3.1.4",
    "loguru>=0.7.2",
    "mail-service>=1.0.5",
    "mailjet-rest>=1.4.0",
    "mercadopago>=2.2.3",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-exporter-jaeger>=1.11.1",
    "opentelemetry-exporter-otlp>=1.27.0",
    "opentelemetry-exporter-zipkin-json>=1.38.0",
    "opentelemetry-instrumentation-asgi>=0.59b0",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "opentelemetry-instrumentation-logging>=0.48b0",
    "opentelemetry-sdk>=1.38.0",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=10.4.0",
    "psycopg[binary,pool]>=3.2",
    "pwdlib[argon2]>=0.2.1",
    "pyjwt>=2.9.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "redis>=5.0.8",
    "resend>=2.10.0",
    "sendgrid>=6.11.0",
    "sentry-sdk[fastapi]>=2.14.0",
    "sqlalchemy[asyncio]>=2.0.34",
    "stripe>=10.10.0",
    "taskiq-faststream>=0.2.0",
    "uvicorn[standard]>=0.30.6",
    "zipp>=3.20.1",
    "google-cloud-recaptcha-enterprise>=1.20.0",
]

[tool.uv]
dev-dependencies = [
    "anyio>=4.4.0",
    "coverage>=7.6.1",
    "factory-boy>=3.3.1",
    "ipdb>=0.13.13",
    "isort>=5.13.2",
    "pyright>=1.1.379",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-env>=1.1.4",
    "pytest-mock>=3.14.0",
    "pytest-ordering>=0.6",
    "pytest-sugar>=1.0.0",
    "pytest-watch>=4.2.0",
    "ruff>=0.6.4",
    "ruff-lsp>=0.0.56",
    "taskipy>=1.13.0",
    "pytest-timeout>=2.3.1",
]

[[tool.uv.index]]
name = "safety"
url = "https://pkgs.safetycli.com/repository/singular-journey/project/fast-ecommerce-back/pypi/simple/"
default = false
[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*", "migrations*", "static*", "legacy*", "contrib*", "hurl_tests*"]

[tool.poetry.group.otel.dependencies]
opentelemetry-distro = "^0.45b0"
opentelemetry-exporter-otlp = "^1.24.0"


[[tool.poetry.source]]
name = "safety"
url = "https://pkgs.safetycli.com/repository/singular-journey/project/fast-ecommerce-back/pypi/simple/"
priority = "primary"

[tool.ruff]
lint.select = ["A", "B", "C", "E", "F", "G", "I", "N", "S", "T", "W", "C4", "ARG", "BLE", "COM", "DTZ", "EM", "ERA", "FAST", "ICN", "ISC", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY"]
lint.fixable = ["A", "B", "C", "E", "F", "G", "N", "S", "T", "W", "ARG", "BLE", "COM", "DTZ", "EM", "ERA", "FAST", "ICN", "ISC", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY"]
lint.unfixable = []
lint.ignore = ["B008"]
src = ["app", "tests"]
# Exclude a variety of commonly ignored directories.
exclude = [ ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "__init__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    ".venv",
    "migrations"
]

[tool.ruff.lint.flake8-annotations]
mypy-init-return = true

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["D104", "N999"]
"/**/tests/*.py" = ["S101", "S105", "S106", "S311", "RUF012", "PLR2004", "B008", 'N806', 'F821', 'N999', 'E501', 'PT011', 'ARG001', 'EM101', 'TRY003', "F823", "F811", "F401", "PLC0415", "ARG002", "I001", "PLR0913", "B017", "PTH123"]
"app/infra/*" = ["B008", "A002","A001", "A003", "RUF100","FBT001", "F811","F401","TRY302"]
"app/*" = ["I001"]
"app/report/*" = ["PLR0913", "TRY003", "EM101"]
"main.py" =  ["ARG001"]
"app/entities/*" = ["F821", "PLR0913"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
known-third-party = ["fastapi"]

[tool.ruff.format]
preview = true
quote-style = 'single'


[tool.pylint.main]
lint.ignore = [ "pyproject.toml", "conftest.py" ]

[tool.pylint.master]
extension-pkg-allow-list = ["pydantic", "rapidjson"]

[tool.pylint.refactoring]
# Maximum number of nested blocks for function / method body
max-nested-blocks = 5

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = ["--strict-markers", "-n0"]
junit_family= "xunit1"
markers = [
  "smoke: main tests to validate system",
  "integrate: integrate tests",
]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore:.*@wait_container_is_ready.*:DeprecationWarning",
    "ignore:.*testcontainers.*:DeprecationWarning",
]

[tool.coverage.run]
omit = [
    "backend/tests/*"
]
concurrency = ["thread", "greenlet"]

[tool.coverage.report]
skip_empty = true


[tool.uv.workspace]
members = ["test"]

[virtualenvs]
path = ".venv"

[dependency-groups]
dev = [
    "faker-file[images,s3]>=0.18.1",
    "fakeredis>=2.32.1",
    "pytest-xdist>=3.6.1",
    "safety>=3.5.2",
    "testcontainers>=4.8.2",
    "towncrier>=24.8.0",
]

[pytest-watch]
ignore = "./tests/endpoints/"

[tool.towncrier]
directory = "changelog.d"
filename = "CHANGELOG.md"
start_string = "<!-- towncrier release notes start -->\n"
underlines = ["", "", ""]
title_format = "## [{version}](https://github.com/twisted/my-project/tree/{version}) - {project_date}"
issue_format = "[#{issue}](https://github.com/twisted/my-project/issues/{issue})"

[[tool.towncrier.type]]
directory = "security"
name = "Security"
showcontent = true

[[tool.towncrier.type]]
directory = "removed"
name = "Removed"
showcontent = true

[[tool.towncrier.type]]
directory = "deprecated"
name = "Deprecated"
showcontent = true

[[tool.towncrier.type]]
directory = "added"
name = "Added"
showcontent = true

[[tool.towncrier.type]]
directory = "changed"
name = "Changed"
showcontent = true

[[tool.towncrier.type]]
directory = "fixed"
name = "Fixed"
showcontent = true

[[tool.towncrier.type]]
directory = "chore"
name = "Chore"
showcontent = true

[[tool.towncrier.type]]
directory = "build"
name = "Build"
showcontent = true

[tool.taskipy.tasks]
install = "uv install"
update = "uv update"
shell = "uv shell"
format = "ruff check . --fix"
lint = "ruff check app/entities && ruff check app/cart/services.py && ruff check app/catalog && ruff check app/coupons && ruff check app/payment/tasks.py && ruff check app/payment/services.py && ruff check app/payment/repository.py && ruff check app/freight && ruff check app/report && ruff check app/user/tasks.py && ruff check app/user/services.py && ruff check app/user/repository.py && ruff check app/inventory/tasks.py && ruff check app/inventory/repository.py && ruff check app/mail && ruff check app/order/tasks.py && ruff check app/order/services.py && ruff check app/order/repository.py && ruff check app/product && ruff check app/settings && ruff check app/infra/endpoints && ruff check tests/ --ignore S101,S105,PLR0913,B017,PTH123"
test = "uv run pytest -s tests/ -x --cov=app -vv"
post_test = "uv run coverage html"
configs = "dynaconf -i src.config.settings list"
upgrade = "uv run alembic upgrade head"
run = "uv run uvicorn main:app --reload --host 0.0.0.0 --port 8001"
migrate = "uv run alembic revision --autogenerate"
task = "uv run taskiq scheduler app.infra.scheduler:scheduler"
safety = "uv run safety scan --continue-on-error || true"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
